Add files via upload
In order to check our qRAM we are giving the input |11> to the address register and then measuring the readout register. For this we have included a classical register in the qRAM circuit. Of course, in the final qRAM there would be no measurement register. The check is to see if the output register gives |1010> with a high probability.